Add config file support for BERT embeddings
Parameter "use_bert" is now included in the config file. The template.libsonnet file
is updated to handle this, using the pretrained model "bert-base-cased". Using the
cased model is advisable for entities recognition. New code is tested on the tasks
"ner", "coref" and "relation" - and it works in these settings.