Add relation propagation functionality
Relation propagation is now added through the usage of the parameter
rel_prop that is fed to the model through the config file.

The function relation_propagation is located in the relation module.

The function update_span_embeddings is located in dygie and is responsible
for updating the embeddings from the relation propagation (and in the
future also from coref propagation).

I had to split up the forward method of relation into two parts to make
this possible. This is because the forward method prunes spans, creates span pair embeddings
and also predicts labels. I splitted up these three functionalities while also keeping the
property that the relation module does not have pointers to dygie.py. Instead, the functions
produce output that is merged into the main forward method (of dygie) through the use of the
function update_span_embeddings.